Update deploy_files.ps1
Updated code to create folder if it does not exist, and to push files to the newly created folder.

+++ deploy_files.ps1
@@ -1,12 +1,11 @@
 <#
 This is a PowerShell script I wrote to push files to a staging environment once they have been merged into the Master Git Branch.
 This is done with a Team City build that will display a unqiue Exit Code if the build fails. 
-To do: Currently if you add a new folder "\images" and it does not already exist in the Staging Environment, then the build chokes.
-Need to add the ability to create a folder if it does not exist, before trying to write a file to it.
 #>
 
 #Initializiations
-$staging="\\server\Path_To_Staging"
+$staging = "\\server\Path_To_Staging"
+$ignoreFileList = New-Object Collections.Generic.List[String]
 $replicatedFiles = New-Object Collections.Generic.List[string]
 
 #Get list of files changed
@@ -16,8 +15,8 @@ $changedFileInfoPath = $args[0]
 $changedFileData = Get-Content $changedFileInfoPath 
 
 #List of files to ignore
-$ignoreFileList = New-Object Collections.Generic.List[String]
 $ignoreFileList.add("deploy_files.ps1")
+$ignoreFileList.add("README.md")
 
 #Iterate through the changed files then add to them to the replication list
 foreach($line in $changedFileData)
@@ -28,31 +27,63 @@ foreach($line in $changedFileData)
 	$replicatedFiles.add($splitLine[0])
 }
 
-#then push these updated files to staging
-#implement try/catch that will return error messages
+#Push updated files to staging and a implement try/catch that will return unique error messages
 try
 {
     foreach ($replicatedFile in $replicatedFiles)
     {
+        #Do not push ignored files
         if ($ignoreFileList -notcontains $replicatedFile)
         {
-            try
+            Write-Output $replicatedFile
+            $pushPath = Join-Path $staging $replicatedFile
+
+            #If Replicated file "contains" a /, then we know it contains a folder (i.e. images/logo.jpg)
+            #Else, it's a straight file copy
+            if ($replicatedFile -like '*/*')
             {
-                $pushPath = Join-Path $staging $replicatedFile
-			    Write-Output $pushPath
-			    Write-Output $replicatedFile
-			    Copy-Item $replicatedFile -Destination $pushPath
+                try
+                {
+                    #If folder in the path does not exist, it will create, and then copy the file
+                    #Else if it exists, it will copy the file
+                    if((Test-Path -Path $pushPath) -eq $false)
+                    {
+                        $folder = $replicatedFile.split('/')[0]
+                        New-Item -Path $testThis -Name $folder -ItemType "directory"
+                        $folderPath = Join-Path $staging $folder
+                        Copy-Item $replicatedFile -Destination $folderPath
+                    } 
+                    else
+                    {
+			            Copy-Item $replicatedFile -Destination $pushPath
+                    }
+                }
+                catch
+                {
+                    Write-Error "Exception: Error creating folder and/or copying filder to folder from Git to the Staging Environment."
+                    Write-Error -Exception $_.Exception
+                    Exit 3 
+                }
             }
-            catch
+            else
             {
-                Write-Error -Exception $_.Exception
-                Exit 1
-            } 
+                try
+                {
+                    Copy-Item $replicatedFile -Destination $pushPath
+                }
+                catch
+                {
+                    Write-Error "Exception: Error copying the file from Git to the Staging Environment."
+                    Write-Error -Exception $_.Exception
+                    Exit 2 
+                }
+            }
         }
     }
 }
 catch
 {
+    Write-Error "An exception has occurred in the main code."
     Write-Error -Exception $_.Exception
-    Exit 2
+    Exit 1
 }
